pyproject.toml
[project]
name = "async-task-processing-engine"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.14"
dependencies = [
    "anyio>=4.12.1",
    "fastapi>=0.129.0",
    "pydantic-settings>=2.13.1",
    "redis>=7.2.0",
    "structlog>=25.5.0",
    "uvicorn[standard]>=0.41.0",
]
